+++ mocapi-server/src/test/java/com/callibrity/mocapi/server/tools/McpToolsServiceTest.java
@@ -73,7 +73,7 @@ class McpToolsServiceTest {
   private McpToolsService service;
 
   private List<CallToolHandler> createHandlers(Object target) {
-    return CallToolHandlers.create(generator, invokerFactory, resolvers, target, s -> s);
+    return CallToolHandlers.discover(target, generator, invokerFactory, resolvers, s -> s);
   }
 
   @BeforeEach